pyproject.toml
[project]
name = "conestoga"
version = "0.1.0"
description = "Oregon Trail-inspired journey simulation with Gemini-powered events"
requires-python = ">=3.12"
dependencies = [
    "google-cloud-storage>=3.8.0",
    "owlrl>=7.1.4",
    "python-dotenv>=1.2.1",
    "rdflib>=7.5.0",
    "pygame-ce>=2.4.0",
    "google-generativeai>=0.3.0",
]
readme = "README.md"
license = {file = "LICENSE"}

[project.scripts]
conestoga = "conestoga.game.runner:main"

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "ruff>=0.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/conestoga"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = []

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

